+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix a bug in cycle detection for nontrivial cycles
+
 ## 0.1.0
 
 - Initial Release